rewrite prepareScriptOptions to be less clever
Making the code much more verbose and comprehensible by converting it to a switch statement. This also solves the previous code's issue where if you omitted the script parameter for a POSTREQUEST script's parameter, it would send a null or blank one to the Data API anyway, which the Data API chokes on. New code otherwise maintains exact function of previous code, possible bugs included.