remove curl escape of entire path
if you curl escape the entire path, as this does, the Data API is incapable of processing any request you send. It will encode the slashes in the path to %2d, which chokes the Data API. This, alongside the rawurlencode of the relevant url parts (via prepareURLpart utility function), should resolve that.